requirements.txt
Django==1.10.4
mysqlclient==1.3.9
Unipath==1.1
djangorestframework==3.5.3